chore(tests): moved cypress setup to webchat package
