+++ packages/webchat/webchat/package.json
@@ -10,13 +10,16 @@
     "watch": "yarn && yarn run -T tsc --build --watch",
     "dev": "yarn run -T parcel test/index.html",
     "copy:scss": "yarn run -T copyfiles -u 1 ./src/**/*.scss dist",
-    "prepublish": "yarn run -T rimraf dist && yarn --immutable && yarn run -T tsc --build && yarn run -T rimraf dist/.tsbuildinfo && yarn copy:scss"
+    "prepublish": "yarn run -T rimraf dist && yarn --immutable && yarn run -T tsc --build && yarn run -T rimraf dist/.tsbuildinfo && yarn copy:scss",
+    "test:e2e": "yarn run -T ts-node test/e2e/cypress/start.ts",
+    "cypress:open": "cypress open --project test/e2e"
   },
   "files": [
     "dist",
     "assets"
   ],
   "devDependencies": {
+    "@botpress/base-test": "0.0.1",
     "@parcel/config-default": "2.2.1",
     "@parcel/transformer-typescript-tsc": "2.2.1",
     "@types/crypto-js": "^4.1.1",
@@ -28,8 +31,10 @@
     "@types/react-dom": "^17.0.16",
     "@types/uuid": "^8.3.4",
     "cypress": "^9.6.0",
+    "portfinder": "^1.0.28",
     "postcss": "^8.4.13",
-    "prop-types": "^15.8.1"
+    "prop-types": "^15.8.1",
+    "typescript": "~4.6.4"
   },
   "dependencies": {
     "@blueprintjs/core": "^3.54.0",
