+++ packages/webchat/webchat/test/e2e/cypress.json
@@ -1,0 +1,12 @@
+{
+  "$schema": "https://on.cypress.io/cypress.schema.json",
+
+  "baseUrl": "http://localhost:1234",
+
+  "integrationFolder": "./",
+
+  "pluginsFile": "cypress/plugins/index.ts",
+  "supportFile": "cypress/support/index.ts",
+
+  "testFiles": "**/*.test.ts"
+}
