+++ packages/webchat/webchat/test/e2e/cypress/plugins/index.ts
@@ -15,7 +15,6 @@
 /**
  * @type {Cypress.PluginConfig}
  */
-// eslint-disable-next-line no-unused-vars
 module.exports = (on, config) => {
   // `on` is used to hook into various events Cypress emits
   // `config` is the resolved Cypress config
