+++ packages/webchat/webchat/test/tsconfig.json
@@ -12,7 +12,11 @@
 
     // be explicit about types included
     // to avoid clashing with Jest types
-    "types": ["cypress"]
+    "types": ["cypress", "node"],
+
+    "paths": {
+      "@botpress/base-test": ["../../../base/test/src/index.ts"]
+    }
   },
   "include": ["../test/**/*", "../src"]
 }
